+++ src/Application/View/Happening/Admin/HappeningDayView.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Proximum\Vimeet\Application\View\Happening\Admin;
+
+use DateTimeInterface;
+
+class HappeningDayView
+{
+    public DateTimeInterface $happeningDay;
+
+    public array $happeningListView;
+
+    public function __construct(DateTimeInterface $happeningDay, array $happeningListView)
+    {
+        $this->happeningDay = $happeningDay;
+        $this->happeningListView = $happeningListView;
+    }
+}